Refactored map node rendering
